requirements.txt
streamlit
pandas
plotly
google-generativeai
openai
python-dotenv
